pyproject.toml
[project]
name = "pubmed-fetcher"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "biopython (>=1.85,<2.0)",
    "pandas (>=2.3.1,<3.0.0)"
]

[tool.poetry]
packages = [{include = "pubmed_fetcher", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
get-papers-list = "pubmed_fetcher.cli:main"
